modify run way

+++ Betweenness/src/main/scala/org/ymt/spark/graphx/betweenness/BetweenCentrality.scala
@@ -13,7 +13,7 @@ object BetweenCentrality extends Serializable{
     val sc = new SparkContext(new SparkConf().setAppName("Intro"))
     val inputPath = args(0)
     val outputPath = args(1)
-    val numPartitions = args(2)
+    val numPartitions = args(2).toInt
 
     // graph loader phase
     val graph = GraphLoader.edgeListFile(sc, inputPath).cache()
