+++ ClusterCoef/src/main/scala/org/ymt/spark/graphx/cluster/ClusterCoef.scala
@@ -13,7 +13,7 @@ object ClusterCoef extends Serializable{
     val sc = new SparkContext(new SparkConf().setAppName("Shortest Path"))
     val inputPath = args(0)
     val outputPath = args(1)
-    val numPartitions = args(2)
+    val numPartitions = args(2).toInt
 
     // graph loader phase
     val graph = GraphLoader.edgeListFile(sc, inputPath).cache()
