+++ Community/src/main/scala/org/ymt/spark/graphx/community/Louvain.scala
@@ -16,7 +16,7 @@ object Louvain extends Serializable{
 
     val inputPath = args(0)
     val outputPath = args(1)
-    val numPartitions = args(2)
+    val numPartitions = args(2).toInt
 
     // graph loader phase
     val graph = GraphLoader.edgeListFile(sc, inputPath).cache()
