+++ Degree/src/main/scala/org/ymt/spark/graphx/degree/DegreeCentrality.scala
@@ -13,7 +13,7 @@ object DegreeCentrality extends Serializable{
     val sc = new SparkContext(new SparkConf().setAppName("Degree Centrality"))
     val inputPath = args(0)
     val outputPath = args(1)
-    val numPartitions = args(2)
+    val numPartitions = args(2).toInt
 
     // graph loader phase
     val graph = GraphLoader.edgeListFile(sc, inputPath).cache()
